[refactor]: sse 응답값 및 swagger 수정
#36
